Internal: Use BigInteger for timestamps (MOB-269). <dl>

Closes MOB-269

See merge request 415-cradle/cradlemobile!109

+++ app/src/main/java/com/cradleVSA/neptune/manager/UrlManager.kt
@@ -2,6 +2,7 @@ package com.cradleVSA.neptune.manager
 
 import android.util.Log
 import com.cradleVSA.neptune.model.Settings
+import java.math.BigInteger
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -109,12 +110,12 @@ class UrlManager @Inject constructor(val settings: Settings) {
      * provides all the updates related to the user since the [currTime] stamp
      */
     @Deprecated("use new sync function", ReplaceWith("getUpdatesNew()"))
-    fun getUpdates(currTime: Long): String = "$base/sync/updates?since=$currTime"
+    fun getUpdates(currTime: BigInteger): String = "$base/sync/updates?since=$currTime"
 
-    fun getPatientsSync(lastSyncTimestamp: Long): String =
+    fun getPatientsSync(lastSyncTimestamp: BigInteger): String =
         "$base/sync/patients?since=$lastSyncTimestamp"
 
-    fun getReadingsSync(lastSyncTimestamp: Long): String =
+    fun getReadingsSync(lastSyncTimestamp: BigInteger): String =
         "$base/sync/readings?since=$lastSyncTimestamp"
 
     fun getUpdatesNew(): String = "$base/sync/updates"
